+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 The template version should stay in sync with the engine version
 
-##
+## 5.0.0 -
 
 - Bugfix with Twine's `hasVisited()`
 - `host.sound.crossfade()`